Add time remaining back for live games
I think this will work, but I need to wait until a game is live before being certain. This fetches all current scores, which includes game clock information, and applies the clock info to games that are in progress. Furthermore, if we aren't able to retrieve the time remaining, we will just display the period number the same way that "Final" or date/time is already displayed.